Move migrations into `UpdateMigrator` module.
