+++ Library/Homebrew/brew.rb
@@ -14,6 +14,8 @@ end
 
 require_relative "global"
 
+require "update_migrator"
+
 begin
   trap("INT", std_trap) # restore default CTRL-C handler
 
@@ -76,7 +78,7 @@ begin
   end
 
   # Migrate LinkedKegs/PinnedKegs if update didn't already do so
-  migrate_legacy_keg_symlinks_if_necessary
+  UpdateMigrator.migrate_legacy_keg_symlinks_if_necessary
 
   # Uninstall old brew-cask if it's still around; we just use the tap now.
   if cmd == "cask" && (HOMEBREW_CELLAR/"brew-cask").exist?
